pyproject.toml
[tool.poetry]
name = "instagram"
version = "0.1.0"
description = ""
authors = ["Sergio Mattei <sergiomattei@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pyexiftool = "^0.5.5"
rich = "^13.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
